safety-quiz/requirements.txt
flask
flask_bootstrap4
sqlalchemy